Internal loan12Mon infor

+++ API/InternalAPI/domain/loanDetail.response.js
@@ -9,8 +9,8 @@ module.exports = function LoanDetailResponse(params, niceSessionKey, sysDtim, wo
 
     this.NICE_SSIN_ID = niceSessionKey;
     this.SEQ = seqs;
-    this.ST_LOAN_VND = parseInt(vnd) ? parseInt(vnd) : 0;
-    this.ST_LOAN_USD = parseInt(usd) ? parseInt(usd) : 0;
+    this.ST_LOAN_VND = parseFloat(vnd) ? parseFloat(vnd) : 0;
+    this.ST_LOAN_USD = parseFloat(usd) ? parseFloat(usd) : 0;
     this.SYS_DTIM = sysDtim;
     this.WORK_ID = workID;
     this.LOAD_DATE;
